LayoutTests/js/codegen-loops-logical-nodes-expected.txt

Tests loop codegen when the condition is a logical node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS while_or_eq() is true
PASS while_or_neq() is true
PASS while_or_less() is true
PASS while_or_lesseq() is true
PASS while_and_eq() is true
PASS while_and_neq() is true
PASS while_and_less() is true
PASS while_and_lesseq() is true
PASS for_or_eq() is true
PASS for_or_neq() is true
PASS for_or_less() is true
PASS for_or_lesseq() is true
PASS for_and_eq() is true
PASS for_and_neq() is true
PASS for_and_less() is true
PASS for_and_lesseq() is true
PASS dowhile_or_eq() is true
PASS dowhile_or_neq() is true
PASS dowhile_or_less() is true
PASS dowhile_or_lesseq() is true
PASS dowhile_and_eq() is true
PASS dowhile_and_neq() is true
PASS dowhile_and_less() is true
PASS dowhile_and_lesseq() is true
PASS while_not_or_eq() is false
PASS while_not_or_neq() is false
PASS while_not_or_less() is false
PASS while_not_or_lesseq() is false
PASS while_not_and_eq() is false
PASS while_not_and_neq() is false
PASS while_not_and_less() is false
PASS while_not_and_lesseq() is false
PASS for_not_or_eq() is false
PASS for_not_or_neq() is false
PASS for_not_or_less() is false
PASS for_not_or_lesseq() is false
PASS for_not_and_eq() is false
PASS for_not_and_neq() is false
PASS for_not_and_less() is false
PASS for_not_and_lesseq() is false
PASS dowhile_not_or_eq() is false
PASS dowhile_not_or_neq() is false
PASS dowhile_not_or_less() is false
PASS dowhile_not_or_lesseq() is false
PASS dowhile_not_and_eq() is false
PASS dowhile_not_and_neq() is false
PASS dowhile_not_and_less() is false
PASS dowhile_not_and_lesseq() is false
PASS float_while_or_eq() is true
PASS float_while_or_neq() is true
PASS float_while_or_less() is true
PASS float_while_or_lesseq() is true
PASS float_while_and_eq() is true
PASS float_while_and_neq() is true
PASS float_while_and_less() is true
PASS float_while_and_lesseq() is true
PASS float_for_or_eq() is true
PASS float_for_or_neq() is true
PASS float_for_or_less() is true
PASS float_for_or_lesseq() is true
PASS float_for_and_eq() is true
PASS float_for_and_neq() is true
PASS float_for_and_less() is true
PASS float_for_and_lesseq() is true
PASS float_dowhile_or_eq() is true
PASS float_dowhile_or_neq() is true
PASS float_dowhile_or_less() is true
PASS float_dowhile_or_lesseq() is true
PASS float_dowhile_and_eq() is true
PASS float_dowhile_and_neq() is true
PASS float_dowhile_and_less() is true
PASS float_dowhile_and_lesseq() is true
PASS float_while_not_or_eq() is false
PASS float_while_not_or_neq() is false
PASS float_while_not_or_less() is false
PASS float_while_not_or_lesseq() is false
PASS float_while_not_and_eq() is false
PASS float_while_not_and_neq() is false
PASS float_while_not_and_less() is false
PASS float_while_not_and_lesseq() is false
PASS float_for_not_or_eq() is false
PASS float_for_not_or_neq() is false
PASS float_for_not_or_less() is false
PASS float_for_not_or_lesseq() is false
PASS float_for_not_and_eq() is false
PASS float_for_not_and_neq() is false
PASS float_for_not_and_less() is false
PASS float_for_not_and_lesseq() is false
PASS float_dowhile_not_or_eq() is false
PASS float_dowhile_not_or_neq() is false
PASS float_dowhile_not_or_less() is false
PASS float_dowhile_not_or_lesseq() is false
PASS float_dowhile_not_and_eq() is false
PASS float_dowhile_not_and_neq() is false
PASS float_dowhile_not_and_less() is false
PASS successfullyParsed is true

TEST COMPLETE
